Extend and refactor cursor tests Description Extending and refactoring the cursor tests. This PR adds 2 new tests to the cursors' tests. One for the `__iter__` and one for `__next__` attributes. Also moving the cursor related tests to a new `test_cursor` file which encapsulates the cursor tests.
