+++ tests/integration/test_cursor.py
@@ -5,10 +5,10 @@ from tests.helpers import IntegrationTestCaseBase, INTEGRATION_TEST_DB
 
 
 @pytest.mark.integration
-class TestCursorNext(IntegrationTestCaseBase):
+class TestCursor(IntegrationTestCaseBase):
     def setup_method(self):
-        super(TestCursorNext, self).setup_method()
-        self.table_name = 'cursor_next'
+        super(TestCursor, self).setup_method()
+        self.table_name = 'test_cursor'
         self.r.table_create(self.table_name).run(self.conn)
         self.documents = [
             {'id': 1, 'name': 'Testing Cursor/Next 1'},
@@ -20,7 +20,7 @@ class TestCursorNext(IntegrationTestCaseBase):
 
     def teardown_method(self):
         self.r.table_drop(self.table_name).run(self.conn)
-        super(TestCursorNext, self).teardown_method()
+        super(TestCursor, self).teardown_method()
 
     def test_get_next_document(self):
         self.r.table(self.table_name).insert(self.documents).run(self.conn)
@@ -58,3 +58,34 @@ class TestCursorNext(IntegrationTestCaseBase):
         with pytest.raises(StopIteration):
             for i in range(0, len(self.documents) + 1):
                 cursor.next()
+
+    def test_for_loop(self):
+        self.r.table(self.table_name).insert(self.documents).run(self.conn)
+
+        documents = list()
+
+        for document in self.r.table(self.table_name).run(self.conn):
+            documents.append(document)
+
+        assert sorted(documents) == sorted(self.documents)
+
+    def test_next(self):
+        self.r.table(self.table_name).insert(self.documents).run(self.conn)
+
+        cursor = self.r.table(self.table_name).run(self.conn)
+
+        assert hasattr(cursor, '__next__')
+
+    def test_iter(self):
+        self.r.table(self.table_name).insert(self.documents).run(self.conn)
+
+        cursor = self.r.table(self.table_name).run(self.conn)
+
+        assert hasattr(cursor, '__iter__')
+
+    def test_close_cursor(self):
+        cursor = self.r.table(self.table_name).run(self.conn)
+        cursor.close()
+
+        assert cursor.conn.is_open()
+        assert isinstance(cursor.error, ReqlCursorEmpty)
